requirements.txt
numpy==1.21.2
Pillow==8.3.1
opencv-python==4.5.3.56